구인 게시판 기능 추가 및 프로젝트 멤버 처리 로직 구현

+++ boxfolio/boxfolio/src/main/java/centum/boxfolio/controller/board/BoardController.java
@@ -70,30 +70,45 @@ public class BoardController {
 
     @GetMapping("/recruit/{boardId}")
     public String recruitBoardPage(@PathVariable Long boardId, Model model, HttpServletRequest request) {
+        HttpSession session = request.getSession(false);
+        Long memberId = Long.parseLong(session.getAttribute(SessionConst.LOGIN_MEMBER).toString());
+
         Recruitment post = boardService.readRecruitPost(boardId);
         List<BoardComment> boardComments = commentService.readCommentsByBoardId(boardId);
+        Boolean applyStatus = boardService.checkApplyStatus(boardId, memberId);
 
         model.addAttribute("recruitPost", post);
         model.addAttribute("comments", boardComments);
         model.addAttribute("boardCommentSaveForm", new BoardCommentSaveForm());
+        model.addAttribute("applyStatus", applyStatus);
         return "board/recruitment_post";
     }
 
     @GetMapping("/free/edit")
     public String freeBoardEditorPage(Model model) {
-        model.addAttribute("freeBoardSaveForm", new FreeBoardSaveForm());
+        FreeBoardSaveForm freeBoardSaveForm = new FreeBoardSaveForm();
+        freeBoardSaveForm.setVisibility("public");
+        freeBoardSaveForm.setCommentAllow(true);
+        model.addAttribute("freeBoardSaveForm", freeBoardSaveForm);
         return "board/free_edit";
     }
 
     @GetMapping("/info/edit")
     public String infoBoardEditorPage(Model model) {
-        model.addAttribute("infoBoardSaveForm", new InfoBoardSaveForm());
+        InfoBoardSaveForm infoBoardSaveForm = new InfoBoardSaveForm();
+        infoBoardSaveForm.setVisibility("public");
+        infoBoardSaveForm.setCommentAllow(true);
+        model.addAttribute("infoBoardSaveForm", infoBoardSaveForm);
         return "board/info_edit";
     }
 
     @GetMapping("/recruit/edit")
     public String recruitBoardEditorPage(Model model) {
-        model.addAttribute("recruitBoardSaveForm", new RecruitBoardSaveForm());
+        RecruitBoardSaveForm recruitBoardSaveForm = new RecruitBoardSaveForm();
+        recruitBoardSaveForm.setVisibility("public");
+        recruitBoardSaveForm.setCommentAllow(true);
+        recruitBoardSaveForm.setMemberTotal(1L);
+        model.addAttribute("recruitBoardSaveForm", recruitBoardSaveForm);
         return "board/recruitment_edit";
     }
 
@@ -105,7 +120,7 @@ public class BoardController {
         }
 
         HttpSession session = request.getSession(false);
-        long memberId = Long.parseLong(session.getAttribute(SessionConst.LOGIN_MEMBER).toString());
+        Long memberId = Long.parseLong(session.getAttribute(SessionConst.LOGIN_MEMBER).toString());
         Free freePost = boardService.createFreePost(freeBoardSaveForm, memberId);
         redirectAttributes.addAttribute("id", freePost.getId());
         return "redirect:/board/free/{id}";
@@ -119,7 +134,7 @@ public class BoardController {
         }
 
         HttpSession session = request.getSession(false);
-        long memberId = Long.parseLong(session.getAttribute(SessionConst.LOGIN_MEMBER).toString());
+        Long memberId = Long.parseLong(session.getAttribute(SessionConst.LOGIN_MEMBER).toString());
         Information infoPost = boardService.createInfoPost(infoBoardSaveForm, memberId);
         redirectAttributes.addAttribute("id", infoPost.getId());
         return "redirect:/board/info/{id}";
@@ -133,29 +148,44 @@ public class BoardController {
         }
 
         HttpSession session = request.getSession(false);
-        long memberId = Long.parseLong(session.getAttribute(SessionConst.LOGIN_MEMBER).toString());
+        Long memberId = Long.parseLong(session.getAttribute(SessionConst.LOGIN_MEMBER).toString());
         Recruitment recruitPost = boardService.createRecruitPost(recruitBoardSaveForm, memberId);
         redirectAttributes.addAttribute("id", recruitPost.getId());
         return "redirect:/board/recruit/{id}";
     }
 
     @GetMapping("/free/modify/{boardId}")
-    public String modifyFreeBoardPage(@PathVariable Long boardId, Model model) {
+    public String modifyFreeBoardPage(@PathVariable Long boardId, Model model, HttpServletRequest request) {
+        HttpSession session = request.getSession(false);
+        Long memberId = Long.parseLong(session.getAttribute(SessionConst.LOGIN_MEMBER).toString());
         Free free = boardService.readFreePost(boardId);
+        if (memberId != free.getMember().getId()) {
+            return "redirect:/board/free";
+        }
         model.addAttribute("freeBoardSaveForm", free.toFreeBoardSaveForm());
         return "board/free_modify";
     }
 
     @GetMapping("/info/modify/{boardId}")
-    public String modifyInfoBoardPage(@PathVariable Long boardId, Model model) {
+    public String modifyInfoBoardPage(@PathVariable Long boardId, Model model, HttpServletRequest request) {
+        HttpSession session = request.getSession(false);
+        Long memberId = Long.parseLong(session.getAttribute(SessionConst.LOGIN_MEMBER).toString());
         Information information = boardService.readInfoPost(boardId);
+        if (memberId != information.getMember().getId()) {
+            return "redirect:/board/info";
+        }
         model.addAttribute("infoBoardSaveForm", information.toInfoBoardSaveForm());
         return "board/info_modify";
     }
 
     @GetMapping("/recruit/modify/{boardId}")
-    public String modifyRecruitBoardPage(@PathVariable Long boardId, Model model) {
+    public String modifyRecruitBoardPage(@PathVariable Long boardId, Model model, HttpServletRequest request) {
+        HttpSession session = request.getSession(false);
+        Long memberId = Long.parseLong(session.getAttribute(SessionConst.LOGIN_MEMBER).toString());
         Recruitment recruitment = boardService.readRecruitPost(boardId);
+        if (memberId != recruitment.getMember().getId()) {
+            return "redirect:/board/recruit";
+        }
         model.addAttribute("recruitBoardSaveForm", recruitment.toRecruitBoardSaveForm());
         return "board/recruitment_modify";
     }
@@ -197,21 +227,69 @@ public class BoardController {
     }
 
     @GetMapping("/free/delete/{boardId}")
-    public String deleteFreeBoard(@PathVariable Long boardId) {
-        boardService.deleteFreeBoard(boardId);
+    public String deleteFreeBoard(@PathVariable Long boardId, HttpServletRequest request) {
+        HttpSession session = request.getSession(false);
+        Long memberId = Long.parseLong(session.getAttribute(SessionConst.LOGIN_MEMBER).toString());
+        Free free = boardService.readFreePost(boardId);
+        if (memberId == free.getMember().getId()) {
+            boardService.deleteFreeBoard(boardId);
+        }
         return "redirect:/board/free";
     }
 
     @GetMapping("/info/delete/{boardId}")
-    public String deleteInfoBoard(@PathVariable Long boardId) {
-        boardService.deleteInfoBoard(boardId);
+    public String deleteInfoBoard(@PathVariable Long boardId, HttpServletRequest request) {
+        HttpSession session = request.getSession(false);
+        Long memberId = Long.parseLong(session.getAttribute(SessionConst.LOGIN_MEMBER).toString());
+        Information information = boardService.readInfoPost(boardId);
+        if (memberId == information.getMember().getId()) {
+            boardService.deleteInfoBoard(boardId);
+        }
         return "redirect:/board/info";
     }
 
     @GetMapping("/recruit/delete/{boardId}")
-    public String deleteRecruitBoard(@PathVariable Long boardId) {
-        boardService.deleteRecruitBoard(boardId);
+    public String deleteRecruitBoard(@PathVariable Long boardId, HttpServletRequest request) {
+        HttpSession session = request.getSession(false);
+        Long memberId = Long.parseLong(session.getAttribute(SessionConst.LOGIN_MEMBER).toString());
+        Recruitment recruitment = boardService.readRecruitPost(boardId);
+        if (memberId == recruitment.getMember().getId()) {
+            boardService.deleteRecruitBoard(boardId);
+        }
         return "redirect:/board/recruit";
+    }
+
+    @GetMapping("/recruit/end/{boardId}")
+    public String endRecruit(@PathVariable Long boardId, HttpServletRequest request, RedirectAttributes redirectAttributes) {
+        HttpSession session = request.getSession(false);
+        Long memberId = Long.parseLong(session.getAttribute(SessionConst.LOGIN_MEMBER).toString());
+        Recruitment recruitment = boardService.readRecruitPost(boardId);
+        if (memberId == recruitment.getMember().getId()) {
+            boardService.endRecruit(boardId);
+        }
+        redirectAttributes.addAttribute("id", recruitment.getId());
+        return "redirect:/board/recruit/{id}";
+    }
+
+    @GetMapping("/recruit/restart/{boardId}")
+    public String restartRecruit(@PathVariable Long boardId, HttpServletRequest request, RedirectAttributes redirectAttributes) {
+        HttpSession session = request.getSession(false);
+        Long memberId = Long.parseLong(session.getAttribute(SessionConst.LOGIN_MEMBER).toString());
+        Recruitment recruitment = boardService.readRecruitPost(boardId);
+        if (memberId == recruitment.getMember().getId()) {
+            boardService.restartRecruit(boardId);
+        }
+        redirectAttributes.addAttribute("id", recruitment.getId());
+        return "redirect:/board/recruit/{id}";
+    }
+
+    @GetMapping("recruit/apply/{boardId}")
+    public String applyRecruit(@PathVariable Long boardId, HttpServletRequest request, RedirectAttributes redirectAttributes) {
+        HttpSession session = request.getSession(false);
+        Long memberId = Long.parseLong(session.getAttribute(SessionConst.LOGIN_MEMBER).toString());
+        boardService.applyRecruit(boardId, memberId);
+        redirectAttributes.addAttribute("id", boardId);
+        return "redirect:/board/recruit/{id}";
     }
 
     @GetMapping("/star/{boardId}")
@@ -254,11 +332,11 @@ public class BoardController {
     @GetMapping("/comment/delete/{commentId}")
     public String deleteComment(@PathVariable Long commentId, HttpServletRequest request) {
         String referer = request.getHeader("referer");
-        HttpSession session = request.getSession(false);
-        Long memberId = Long.parseLong(session.getAttribute(SessionConst.LOGIN_MEMBER).toString());
         if (referer != null) {
             commentService.deleteComment(commentId);
         }
+        HttpSession session = request.getSession(false);
+        Long memberId = Long.parseLong(session.getAttribute(SessionConst.LOGIN_MEMBER).toString());
 
         return "redirect:" + referer;
     }
